train_model_function/requirements.txt
# 11.train_model_function/requirements.txt
azure-functions
requests
python-dotenv 